py-lib/pyproject.toml
[project]
name = "reg23-experiments"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohappyeyeballs==2.4.4",
    "aiohttp==3.11.11",
    "aiosignal==1.3.2",
    "anyio==4.8.0",
    "argon2-cffi==23.1.0",
    "argon2-cffi-bindings==21.2.0",
    "arrow==1.3.0",
    "asttokens==3.0.0",
    "attrs==24.3.0",
    "beautifulsoup4==4.12.3",
    "bleach==6.2.0",
    "certifi==2024.12.14",
    "cffi==1.17.1",
    "charset-normalizer==3.4.1",
    "click==8.1.8",
    "cmocean==4.0.3",
    "colorcet==3.1.0",
    "comm==0.2.2",
    "contourpy==1.3.1",
    "cycler==0.12.1",
    "decorator==5.1.1",
    "defusedxml==0.7.1",
    "deprecated==1.2.16",
    "einops==0.8.0",
    "executing==2.2.0",
    "fastcore==1.7.28",
    "fastjsonschema==2.21.1",
    "filelock==3.16.1",
    "fonttools==4.55.3",
    "fqdn==1.5.1",
    "frozenlist==1.5.0",
    "fsspec==2024.12.0",
    "glob2==0.7",
    "huggingface-hub==0.27.1",
    "humanize==4.11.0",
    "idna==3.10",
    "imageio==2.37.0",
    "ipython==8.31.0",
    "ipywidgets==8.1.5",
    "isoduration==20.11.0",
    "jedi==0.19.2",
    "jinja2==3.1.5",
    "jsonpointer==3.0.0",
    "jsonschema==4.23.0",
    "jsonschema-specifications==2024.10.1",
    "jupyter-client==8.6.3",
    "jupyter-core==5.7.2",
    "jupyter-events==0.11.0",
    "jupyter-server==2.15.0",
    "jupyter-server-proxy==4.4.0",
    "jupyter-server-terminals==0.5.3",
    "jupyterlab-pygments==0.3.0",
    "jupyterlab-widgets==3.0.13",
    "kiwisolver==1.4.8",
    "kornia==0.8.0",
    "kornia-rs==0.1.8",
    "magicgui>=0.10.0",
    "markdown-it-py==3.0.0",
    "markupsafe==3.0.2",
    "matplotlib==3.10.0",
    "matplotlib-inline==0.1.7",
    "mdurl==0.1.2",
    "meshio==5.3.5",
    "mistune==3.1.0",
    "more-itertools==10.6.0",
    "mpmath==1.3.0",
    "msgpack==1.1.0",
    "multidict==6.1.0",
    "nbclient==0.10.2",
    "nbconvert==7.16.5",
    "nbformat==5.10.4",
    "nest-asyncio==1.6.0",
    "networkx==3.4.2",
    "nibabel==5.3.2",
    "numpy==2.2.1",
    "overrides==7.7.0",
    "packaging==24.2",
    "pandas==2.2.3",
    "pandocfilters==1.5.1",
    "parso==0.8.4",
    "pexpect==4.9.0",
    "pillow==11.1.0",
    "platformdirs==4.3.6",
    "plotly==5.24.1",
    "pooch==1.8.2",
    "prometheus-client==0.21.1",
    "prompt-toolkit==3.0.50",
    "propcache==0.2.1",
    "ptyprocess==0.7.0",
    "pure-eval==0.2.3",
    "pycparser==2.22",
    "pydicom==3.0.1",
    "pygments==2.19.1",
    "pynrrd==1.1.1",
    "pyparsing==3.2.1",
    "pyqt6>=6.8.0",
    "pyswarms>=1.3.0",
    "pytest>=8.3.4",
    "python-dateutil==2.9.0.post0",
    "python-json-logger==3.2.1",
    "pytz==2024.2",
    "pyvista==0.44.2",
    "pyyaml==6.0.2",
    "pyzmq==26.2.0",
    "referencing==0.36.1",
    "requests==2.32.3",
    "rfc3339-validator==0.1.4",
    "rfc3986-validator==0.1.1",
    "rich==13.9.4",
    "rpds-py==0.22.3",
    "safetensors==0.5.2",
    "scipy==1.15.1",
    "scooby==0.10.0",
    "seaborn==0.13.2",
    "send2trash==1.8.3",
    "setuptools==75.8.0",
    "shellingham==1.5.4",
    "simpervisor==1.0.0",
    "simpleitk==2.4.1",
    "six==1.17.0",
    "sniffio==1.3.1",
    "soupsieve==2.6",
    "stack-data==0.6.3",
    "tenacity==9.0.0",
    "terminado==0.18.1",
    "timm==1.0.14",
    "tinycss2==1.4.0",
    "torchio==0.20.13",
    "tornado==6.4.2",
    "tqdm==4.67.1",
    "traitlets==5.14.3",
    "trame==3.7.6",
    "trame-client==3.5.2",
    "trame-server==3.3.0",
    "trame-vtk==2.8.14",
    "trame-vuetify==2.8.1",
    "typer==0.15.1",
    "types-python-dateutil==2.9.0.20241206",
    "typing-extensions==4.12.2",
    "tzdata==2025.1",
    "uri-template==1.3.0",
    "urllib3==2.3.0",
    "vtk==9.3.1",
    "wcwidth==0.2.13",
    "webcolors==24.11.1",
    "webencodings==0.5.1",
    "websocket-client==1.8.0",
    "widgetsnbextension==4.0.13",
    "wrapt==1.17.2",
    "wslink==2.2.2",
    "yarl==1.18.3",
    "objgraph>=3.6.2",
    "pylibjpeg>=2.0.1",
    "pylibjpeg-libjpeg>=2.3.0",
    "torchviz>=0.0.3",
    "napari[pyqt6, optional]>=0.6.6; sys_platform == 'darwin'",
    "napari[all]>=0.6.6; sys_platform != 'darwin'",
    "scipy-stubs==1.16.3.3",
    "graphviz>=0.21",
    "pyarrow>=23.0.0",
    "nvidia-ml-py>=13.590.48",
]

# sets of optional dependencies (distinct from the normal dependencies above) are defined
[project.optional-dependencies]
cpu = [
    "reg23[cpu]", # the option in the square brackets is a group of optional dependencies
]
cuda = [
    "reg23[cuda]",
]

[dependency-groups]
docs = [
    "pdoc>=16.0.0",
]

[tool.uv]
# the project is re-built when any of the following globbed files are modified
cache-keys = [
    { file = "pyproject.toml" },
]
# conflicts between optional dependencies are stated so that uv doesn't try to install them by default and complain
conflicts = [
    [
        { extra = "cpu" },
        { extra = "cuda" },
    ],
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
